soma_ai/requirements.txt
django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
drf-spectacular==0.27.*
django-redis==5.4.*
celery==5.3.*
redis==5.0.*
psycopg2-binary==2.9.*
python-decouple==3.8
django-storages==1.14.*
dj-database-url==2.1.*
anthropic==0.28.*
openai==1.30.*
pytest-django==4.8.*
factory-boy==3.3.*
coverage==7.5.*
pytest-cov==5.0.*
